LayoutTests/http/tests/pointer-lock/iframe-sandboxed-nested-allow-pointer-lock-expected.txt

Test nested sandboxed iframes with allow-pointer-lock allow pointer lock.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS message is "inner-iframe.html onwebkitpointerlockchange, document.webkitPointerLockElement = [object HTMLBodyElement]"
PASS successfullyParsed is true

TEST COMPLETE
